backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn==1.5.0
pandas
numpy
python-dotenv
groq
reportlab
firebase-admin
pydantic
